requirements.txt
python-dotenv
selenium
pandas
loguru